Sample network edge list format for this program
This is an edge list for the UNdirected 'Zachary's Karate Club' network.  This is an example of how network data must be formatted for input into this code. The code used to input data into the program is not sophisticated and requires strict adherence to this formatting.

Vertex numbering must start at zero and no numbers can be skipped. For example, if there are N vertices, they should be numbered {0,1,2,....,(N-1)} with no missing integers in the set. If vertex 0<=L<=(N-1) is missing from the edge list, the program may accept the data but it will load the edges to incorrect positions and produce corrupted results.

Each edge should appear twice in the edge list. If vertices 0 and 1 are connected then there should be two rows in the edge list: 0 1 and 1 0. 

The edge list should be sorted least-to-greatest first on the left column and then likewise on the right column.